feat: set preferred video codec via a query param
