+++ packages/react-sdk/src/components/Debug/useIsDebugMode.ts
@@ -1,0 +1,18 @@
+import { useMemo } from 'react';
+
+const useQueryParams = () => {
+  return useMemo(() => new URLSearchParams(window.location.search), []);
+};
+
+/**
+ * Internal purpose hook. Enables certain dev mode tools.
+ */
+export const useIsDebugMode = () => {
+  const params = useQueryParams();
+  return !!params.get('debug');
+};
+
+export const useDebugPreferredVideoCodec = () => {
+  const params = useQueryParams();
+  return params.get('video_codec');
+};
